refactor: use lightweight Alpaca coin logo
